auth and idk

+++ app/Http/Controllers/ProductController.php
@@ -4,52 +4,49 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\StoreProductRequest;
 use App\Http\Requests\UpdateProductRequest;
-use App\Models\Product;
 use App\Http\Resources\ProductResource;
-use App\Models\Company;
+use App\Models\Product;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class ProductController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     */
     public function index()
     {
-        return ProductResource::collection(Product::paginate(5));
+        $products = Auth::user()->company->products()->paginate(5);
+
+        return ProductResource::collection($products);
     }
 
-    /**
-     * Store a newly created resource in storage.
-     */
     public function store(StoreProductRequest $request)
     {
-        $product = Product::create($request->validated());
+        $user = Auth::user();
+        $company = $user->company;
+        $product = $company->products()->create($request->validated());
+
         return new ProductResource($product);
     }
 
-    /**
-     * Display the specified resource.
-     */
     public function show(Product $product)
     {
+        Gate::authorize('viewOrModify', $product);
+
         return new ProductResource($product);
     }
 
-    /**
-     * Update the specified resource in storage.
-     */
     public function update(UpdateProductRequest $request, Product $product)
     {
+        Gate::authorize('viewOrModify', $product);
         $product->update($request->validated());
+
         return new ProductResource($product);
     }
 
-    /**
-     * Remove the specified resource from storage.
-     */
     public function destroy(Product $product)
     {
+        Gate::authorize('viewOrModify', $product);
         $product->delete();
+
         return response()->noContent();
     }
 }
